requirements.txt
Click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
psycopg2
python-dotenv